move `location` module to polonius legacy module
this is specific to the old datalog implementation and wasn't noticed in
the previous module move
