+++ compiler/rustc_borrowck/src/consumers.rs
@@ -9,10 +9,10 @@ pub use super::borrow_set::{BorrowData, BorrowSet, TwoPhaseActivation};
 pub use super::constraints::OutlivesConstraint;
 pub use super::dataflow::{BorrowIndex, Borrows, calculate_borrows_out_of_scope_at_location};
 pub use super::facts::{AllFacts as PoloniusInput, PoloniusRegionVid, RustcFacts};
-pub use super::location::{LocationTable, RichLocation};
 pub use super::nll::PoloniusOutput;
 pub use super::place_ext::PlaceExt;
 pub use super::places_conflict::{PlaceConflictBias, places_conflict};
+pub use super::polonius::legacy::{LocationTable, RichLocation};
 pub use super::region_infer::RegionInferenceContext;
 
 /// Options determining the output behavior of [`get_body_with_borrowck_facts`].
