+++ compiler/rustc_borrowck/src/facts.rs
@@ -11,7 +11,7 @@ use rustc_middle::ty::{RegionVid, TyCtxt};
 use rustc_mir_dataflow::move_paths::MovePathIndex;
 
 use crate::BorrowIndex;
-use crate::location::{LocationIndex, LocationTable};
+use crate::polonius::legacy::{LocationIndex, LocationTable};
 
 #[derive(Copy, Clone, Debug)]
 pub struct RustcFacts;
