+++ compiler/rustc_borrowck/src/lib.rs
@@ -53,11 +53,11 @@ use crate::borrow_set::{BorrowData, BorrowSet};
 use crate::consumers::{BodyWithBorrowckFacts, ConsumerOptions};
 use crate::dataflow::{BorrowIndex, Borrowck, BorrowckDomain, Borrows};
 use crate::diagnostics::{AccessKind, IllegalMoveOriginKind, MoveError, RegionName};
-use crate::location::LocationTable;
 use crate::nll::PoloniusOutput;
 use crate::path_utils::*;
 use crate::place_ext::PlaceExt;
 use crate::places_conflict::{PlaceConflictBias, places_conflict};
+use crate::polonius::legacy::LocationTable;
 use crate::prefixes::PrefixSet;
 use crate::region_infer::RegionInferenceContext;
 use crate::renumber::RegionCtxt;
@@ -70,7 +70,6 @@ mod dataflow;
 mod def_use;
 mod diagnostics;
 mod facts;
-mod location;
 mod member_constraints;
 mod nll;
 mod path_utils;
