+++ compiler/rustc_borrowck/src/nll.rs
@@ -28,8 +28,8 @@ use crate::borrow_set::BorrowSet;
 use crate::consumers::ConsumerOptions;
 use crate::diagnostics::RegionErrors;
 use crate::facts::{AllFacts, AllFactsExt, RustcFacts};
-use crate::location::LocationTable;
 use crate::polonius::LocalizedOutlivesConstraintSet;
+use crate::polonius::legacy::LocationTable;
 use crate::region_infer::RegionInferenceContext;
 use crate::type_check::{self, MirTypeckResults};
 use crate::universal_regions::UniversalRegions;
