+++ compiler/rustc_borrowck/src/polonius/legacy/loan_invalidations.rs
@@ -9,9 +9,9 @@ use rustc_middle::mir::{
 use rustc_middle::ty::TyCtxt;
 use tracing::debug;
 
+use super::LocationTable;
 use crate::borrow_set::BorrowSet;
 use crate::facts::AllFacts;
-use crate::location::LocationTable;
 use crate::path_utils::*;
 use crate::{
     AccessDepth, Activation, ArtificialField, BorrowIndex, Deep, LocalMutationIsAllowed, Read,
