+++ compiler/rustc_borrowck/src/polonius/legacy/loan_kills.rs
@@ -6,9 +6,9 @@ use rustc_middle::mir::{
 use rustc_middle::ty::TyCtxt;
 use tracing::debug;
 
+use super::LocationTable;
 use crate::borrow_set::BorrowSet;
 use crate::facts::AllFacts;
-use crate::location::LocationTable;
 use crate::places_conflict;
 
 /// Emit `loan_killed_at` and `cfg_edge` facts at the same time.
