+++ compiler/rustc_borrowck/src/polonius/legacy/mod.rs
@@ -14,7 +14,6 @@ use tracing::debug;
 use crate::borrow_set::BorrowSet;
 use crate::constraints::OutlivesConstraint;
 use crate::facts::{AllFacts, PoloniusRegionVid};
-use crate::location::LocationTable;
 use crate::type_check::MirTypeckRegionConstraints;
 use crate::type_check::free_region_relations::UniversalRegionRelations;
 use crate::universal_regions::UniversalRegions;
@@ -22,6 +21,8 @@ use crate::universal_regions::UniversalRegions;
 mod accesses;
 mod loan_invalidations;
 mod loan_kills;
+mod location;
+pub use self::location::*;
 
 /// When requested, emit most of the facts needed by polonius:
 /// - moves and assignments
