+++ compiler/rustc_borrowck/src/type_check/mod.rs
@@ -48,9 +48,9 @@ use crate::borrow_set::BorrowSet;
 use crate::constraints::{OutlivesConstraint, OutlivesConstraintSet};
 use crate::diagnostics::UniverseInfo;
 use crate::facts::AllFacts;
-use crate::location::LocationTable;
 use crate::member_constraints::MemberConstraintSet;
 use crate::polonius::PoloniusContext;
+use crate::polonius::legacy::LocationTable;
 use crate::region_infer::TypeTest;
 use crate::region_infer::values::{LivenessValues, PlaceholderIndex, PlaceholderIndices};
 use crate::renumber::RegionCtxt;
